flush out statsdb interface fork, and noms implementation
