--- go/libraries/doltcore/sqle/statspro/database.go
@@ -1,15 +1,0 @@
-package statspro
-
-import "github.com/dolthub/go-mysql-server/sql"
-
-// database needs to read and write statistics
-// also need to delete/add specific chunks
-type Database interface {
-	ListStatQuals() []sql.StatQualifier
-	Load(string) error
-	HasStat()
-	GetStat()
-	GetAllStats()
-	PutChunk()
-	DeleteChunk()
-}
